Feat: Redis 설정을 위한 RedisProperties 클래스 추가
