+++ src/main/java/com/beyond/jellyorder/common/auth/RedisProperties.java
@@ -1,0 +1,13 @@
+package com.beyond.jellyorder.common.auth;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "spring.redis")
+@Getter
+@Setter
+public class RedisProperties {
+    private String host;
+    private int port;
+}
