--- src/main/java/com/beyond/jellyorder/config/RedisConfig.java
@@ -1,141 +1,0 @@
-package com.beyond.jellyorder.config;
-
-import com.beyond.jellyorder.domain.sseRequest.service.SseAlarmService;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.PatternTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-public class RedisConfig {
-    @Value("${spring.redis.host}")
-    private String host;
-    @Value("${spring.redis.port}")
-    private int port;
-
-    /* RefreshToken Redis에 저장하기 위한 Factory, 0번에 설정하였으나 추후 변경 가능 합니다! */
-    @Bean
-    @Qualifier("rtInventory")
-    public RedisConnectionFactory redisConnectionFactory() {
-        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration();
-        configuration.setHostName(host);
-        configuration.setPort(port);
-        configuration.setDatabase(0);
-
-        return new LettuceConnectionFactory(configuration);
-    }
-
-    /* RefreshToken RedisTemplate */
-    @Bean
-    @Qualifier("rtInventory")
-    public RedisTemplate<String, Object> redisTemplate(@Qualifier("rtInventory") RedisConnectionFactory redisConnectionFactory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(redisConnectionFactory);
-
-        // 직렬화 설정
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(new Jackson2JsonRedisSerializer<>(Object.class));
-        return template;
-    }
-
-    // 개수 관리 redis
-    @Bean
-    @Qualifier("stockInventory")
-    public RedisConnectionFactory stockConnectionFactory() {
-        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration();
-        configuration.setHostName(host);
-        configuration.setPort(port);
-        configuration.setDatabase(1);
-
-        return (new LettuceConnectionFactory(configuration));
-    }
-
-    @Bean
-    @Qualifier("stockInventory")
-    public RedisTemplate<String, Object> stockTemplate(@Qualifier("stockInventory") RedisConnectionFactory redisConnectionFactory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(redisConnectionFactory);
-
-        // 직렬화 설정
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(new Jackson2JsonRedisSerializer<>(Object.class));
-        return template;
-    }
-
-    @Bean
-    @Qualifier("sseRtInventory")
-    public RedisConnectionFactory sseRedisConnectionFactory() {
-        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration();
-        configuration.setHostName(host);
-        configuration.setPort(port);
-        configuration.setDatabase(0);
-
-        return (new LettuceConnectionFactory(configuration));
-    }
-
-    // RedisTemplate Bean : 데이터를 저장하고 조회하는 주요 도구
-    @Bean
-    @Qualifier("sseRtInventory")
-    public RedisTemplate<String, Object> sseRedisTemplate(@Qualifier("sseRtInventory") RedisConnectionFactory factory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(factory);
-
-        // 직렬화 설정
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(new Jackson2JsonRedisSerializer<>(Object.class));
-        return template;
-    }
-
-    // Redis 채널로부터 수신된 메시지를 처리할 리스너 등록
-    @Bean
-    public MessageListenerAdapter messageListenerAdapter(SseAlarmService sseAlarmService) {
-        return new MessageListenerAdapter(sseAlarmService, "onMessage"); // onMessage() 호출
-    }
-
-    // Redis PubSub 리스너 컨테이너 등록
-    @Bean
-    public RedisMessageListenerContainer redisMessageListenerContainer(
-            RedisConnectionFactory redisConnectionFactory,
-            MessageListenerAdapter listenerAdapter
-    ) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(redisConnectionFactory);
-        container.addMessageListener(listenerAdapter, new PatternTopic("order-channel")); // 원하는 채널명
-        return container;
-    }
-
-    // redis pub/sub을 위한 연결객체 생성
-    @Bean
-    @Qualifier("ssePubSub")
-    public RedisConnectionFactory sseFactory() {
-        RedisStandaloneConfiguration configuration = new RedisStandaloneConfiguration();
-        configuration.setHostName(host);
-        configuration.setPort(port);
-
-        return (new LettuceConnectionFactory(configuration));
-    }
-
-    // ssePubSub 템플릿 객체
-    @Bean
-    @Qualifier("ssePubSub")
-    public RedisTemplate<String, Object> ssePubSubRedisTemplate(@Qualifier("ssePubSub") RedisConnectionFactory redisConnectionFactory) {
-        RedisTemplate<String, Object> template = new RedisTemplate<>();
-        template.setConnectionFactory(redisConnectionFactory);
-
-        // 직렬화 설정
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(new Jackson2JsonRedisSerializer<>(Object.class));
-
-        return template;
-    }
-
-}
